requirements.txt
altair==4.1.0
amqp==5.0.6
appdirs==1.4.4
arabic-reshaper==2.1.3
argh==0.26.2
asgiref==3.3.1
astor==0.8.1
atomicwrites==1.4.0
backports.weakref==1.0.post1
base58==2.1.0
billiard==3.6.4.0
bkcharts==0.2
black==19.10b0
blinker==1.4
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
celery==5.1.0
certifi==2020.12.5
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0

clyent==1.2.2

commonmark==0.9.1
comtypes==1.1.9

contextlib2==0.6.0.post1

cycler==0.10.0

cytoolz==0.11.0

dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-filter==2.4.0

entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0

future==0.18.2

gitdb==4.0.7
GitPython==3.1.17

gunicorn==20.1.0

h2==3.2.0
h5py==2.10.0
HeapDict==1.0.1
hpack==3.0.0

hyperframe==5.2.0

jdcal==1.4.1

json5==0.9.5


keyrings.alt==4.0.2

kombu==5.1.0





multipledispatch==0.6.0
multitasking==0.0.9



oauthlib==3.1.1
olefile==0.46

parso==0.7.0

pathspec==0.7.0
patsy==0.5.1
pep8==1.7.1

pkginfo==1.7.0
ply==3.11

protobuf==3.17.1

psycopg2==2.8.6

pyarrow==4.0.0

pycosat==0.6.3



PyJWT==2.1.0

PyPDF2==1.26.0

pyreadline==2.1

pytest==6.2.3
python-bidi==0.4.2

python-decouple==3.4

python3-openid==3.2.0

pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7

pyzmq==20.0.0
QDarkStyle==2.8.1

QtPy==1.9.0
reportlab==3.5.68

requests-oauthlib==1.3.0

scikit-image==0.18.1

serverfiles==0.3.1
simplegeneric==0.8.1

sip==4.19.13

sklearn==0.0
smmap==4.0.0

spicy==0.16.0

sqlparse==0.4.1
statsmodels==0.12.2
streamlit==0.82.0

tables==3.6.1




whitenoise==5.2.0
widgetsnbextension==3.5.1

win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xhtml2pdf==0.2.5

xlwings==0.23.0
xlwt==1.3.0
xmltodict==0.12.0


zict==2.0.0

zope.event==4.5.0

